管理员 角色 增删改查

+++ shop_common/src/main/java/com/bt/pojo/DtsStorage.java
@@ -47,7 +47,7 @@ public class DtsStorage implements Serializable {
     /**
      * 文件大小
      */
-    private Integer size;
+    private long size;
 
     /**
      * 文件访问链接
